Correct Observer debug name to point to constructing frame
Having been moved into a getName() call, the function failed to serve
its purpose of pointing to a location in user source for easier
debugging.

Determination of the stack frame has also been improved to search for
the first non-constructor, in order to support subclasses of Observable.
This functionality has been tested in the browser and VM.

Fixes #361

MR: constructing_frame2